+++ database/migrations/2019_12_14_000001_create_personal_access_tokens_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (app()->runningUnitTests()) {
+            return ;
+        }
         Schema::create('personal_access_tokens', function (Blueprint $table) {
             $table->id();
             $table->morphs('tokenable');
@@ -28,6 +31,9 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (app()->runningUnitTests()) {
+            return ;
+        }
         Schema::dropIfExists('personal_access_tokens');
     }
 };
